+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Improvements
 - Arch
     - repositories/AUR search time (=~ -70%)
+- Core
+    - saving settings time (=~ -11%)
+    
 ### Fixes
 - Arch
     - crashing when information of a given package is not available